Будем знакомы

+++ routes/api.php
@@ -23,43 +23,43 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::mfilmdleware('auth:sanctum')->get('/user', function (Request $request) {
+Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
 Route::post('/register', [RegisterController::class, 'register']);
 Route::post('/login', [AuthController::class, 'login']);
 Route::post('/logout', [AuthController::class, 'logout']);
-Route::get('/favorite', [FavoriteController::class, 'index']);
+Route::get('/favorite', [FavoriteController::class, 'index'])->middleware('auth:sanctum');
 
-Route::prefix('/user')->group(function () {
+Route::prefix('/user')->middleware('auth:sanctum')->group(function () {
     Route::get('/', [UserController::class, 'show']);
     Route::patch('/', [UserController::class, 'update']);
 });
 
 Route::prefix('/films')->group(function () {
-    Route::post('/{film}/favorite/', [FavoriteController::class, 'store']);
-    Route::delete('/{film}/favorite/', [FavoriteController::class, 'destroy']);
+    Route::post('/{film}/favorite/', [FavoriteController::class, 'store'])->middleware('auth:sanctum');
+    Route::delete('/{film}/favorite/', [FavoriteController::class, 'destroy'])->middleware('auth:sanctum');
     Route::get('/', [FilmController::class, 'index']);
     Route::get('/{film}', [FilmController::class, 'show']);
     Route::post('/', [FilmController::class, 'store']);
-    Route::patch('/{film}', [FilmController::class, 'update']);
+    Route::patch('/{film}', [FilmController::class, 'update'])->middleware('auth:sanctum');
     Route::get('/{film}/similar', [SimilarController::class, 'index']);
     Route::get('/{film}/comments', [CommentController::class, 'index']);
-    Route::post('/{film}/comments', [CommentController::class, 'store']);
+    Route::post('/{film}/comments', [CommentController::class, 'store'])->middleware('auth:sanctum');
 });
 
 Route::prefix('/genres')->group(function () {
     Route::get('/', [GenreController::class, 'index']);
-    Route::patch('/{genre}', [GenreController::class, 'update']);
+    Route::patch('/{genre}', [GenreController::class, 'update'])->middleware('auth:sanctum');
 });
 
-Route::prefix('/comments')->group(function () {
+Route::prefix('/comments')->middleware('auth:sanctum')->group(function () {
     Route::patch('/{comment}', [CommentController::class, 'update']);
     Route::delete('/{comment}', [CommentController::class, 'destroy']);
 });
 
 Route::prefix('/promo')->group(function () {
     Route::get('/', [PromoController::class, 'index']);
-    Route::post('/{film}', [PromoController::class, 'store']);
+    Route::post('/{film}', [PromoController::class, 'store'])->middleware('auth:sanctum');
 });
